Combined PDF to Word/Excel/PowerPoint samples into one
