--- Samples/PDF2ExcelTest/PYTHON/RunTest.bat
@@ -1,8 +1,0 @@
-@echo off
-setlocal
-call ..\..\py_init.bat
-if %ERRORLEVEL% NEQ 0 goto EOF
-set TEST_NAME=PDF2ExcelTest
-python.exe -u %TEST_NAME%.py
-:EOF
-endlocal
