+++ Samples/PDF2OfficeTest/PHP/RunTest.bat
@@ -1,6 +1,6 @@
 @echo off
 setlocal
-set TEST_NAME=PDF2ExcelTest
+set TEST_NAME=PDF2OfficeTest
 set PATH=..\..\..\PDFNetC\Lib;%PATH%
 php.exe %TEST_NAME%.php
 endlocal
