+++ Samples/PDF2OfficeTest/PHP/RunTest.sh
@@ -1,4 +1,4 @@
 #!/bin/sh
-TEST_NAME=PDF2ExcelTest
+TEST_NAME=PDF2OfficeTest
 export LD_LIBRARY_PATH=../../../PDFNetC/Lib
-ruby $TEST_NAME.rb
+php $TEST_NAME.php
