+++ Samples/PDF2OfficeTest/PYTHON/RunTest.bat
@@ -2,7 +2,7 @@
 setlocal
 call ..\..\py_init.bat
 if %ERRORLEVEL% NEQ 0 goto EOF
-set TEST_NAME=PDF2WordTest
+set TEST_NAME=PDF2OfficeTest
 python.exe -u %TEST_NAME%.py
 :EOF
 endlocal
