+++ Samples/PDF2OfficeTest/PYTHON/RunTest.sh
@@ -1,5 +1,5 @@
 #!/bin/sh
-TEST_NAME=PDF2ExcelTest
+TEST_NAME=PDF2OfficeTest
 export LD_LIBRARY_PATH=../../../PDFNetC/Lib
 . ../../py_init.sh
 $python_exe -u $TEST_NAME.py
