+++ Samples/PDF2OfficeTest/RUBY/PDF2OfficeTest.rb
@@ -11,7 +11,7 @@ $stdout.sync = true
 
 #---------------------------------------------------------------------------------------
 # The following sample illustrates how to use the PDF.Convert utility class to convert 
-# documents and files to PowerPoint.
+# documents and files to Word, Excel and PowerPoint.
 #
 # The Structured Output module is an optional PDFNet Add-on that can be used to convert PDF
 # and other documents into Word, Excel, PowerPoint and HTML format.
@@ -49,6 +49,74 @@ def main()
 	#-----------------------------------------------------------------------------------
 
 	begin
+		# Convert PDF document to Word
+		puts "Converting PDF to Word"
+
+		$outputFile = $outputPath + "paragraphs_and_tables.docx"
+
+		Convert.ToWord($inputPath + "paragraphs_and_tables.pdf", $outputFile)
+
+		puts "Result saved in " + $outputFile
+	rescue => error
+		puts "Unable to convert PDF document to Word, error: " + error.message
+	end
+
+	#-----------------------------------------------------------------------------------
+	
+	begin
+		# Convert PDF document to Word with options
+		puts "Converting PDF to Word with options"
+
+		$outputFile = $outputPath + "paragraphs_and_tables_first_page.docx"
+
+		$wordOutputOptions = Convert::WordOutputOptions.new()
+
+		# Convert only the first page
+		$wordOutputOptions.SetPages(1, 1);
+
+		Convert.ToWord($inputPath + "paragraphs_and_tables.pdf", $outputFile, $wordOutputOptions)
+		puts "Result saved in " + $outputFile
+	rescue => error
+		puts "Unable to convert PDF document to Word, error: " + error.message
+	end
+
+	#-----------------------------------------------------------------------------------
+
+	begin
+		# Convert PDF document to Excel
+		puts "Converting PDF to Excel"
+
+		$outputFile = $outputPath + "paragraphs_and_tables.xlsx"
+
+		Convert.ToExcel($inputPath + "paragraphs_and_tables.pdf", $outputFile)
+
+		puts "Result saved in " + $outputFile
+	rescue => error
+		puts "Unable to convert PDF document to Excel, error: " + error.message
+	end
+
+	#-----------------------------------------------------------------------------------
+	
+	begin
+		# Convert PDF document to Excel with options
+		puts "Converting PDF to Excel with options"
+
+		$outputFile = $outputPath + "paragraphs_and_tables_second_page.xlsx"
+
+		$excelOutputOptions = Convert::ExcelOutputOptions.new()
+
+		# Convert only the second page
+		$excelOutputOptions.SetPages(2, 2);
+
+		Convert.ToExcel($inputPath + "paragraphs_and_tables.pdf", $outputFile, $excelOutputOptions)
+		puts "Result saved in " + $outputFile
+	rescue => error
+		puts "Unable to convert PDF document to Excel, error: " + error.message
+	end
+
+	#-----------------------------------------------------------------------------------
+
+	begin
 		# Convert PDF document to PowerPoint
 		puts "Converting PDF to PowerPoint"
 
@@ -81,6 +149,7 @@ def main()
 	end
 
 	#-----------------------------------------------------------------------------------
+
 	PDFNet.Terminate
 	puts "Done."
 end
