--- Samples/PDF2PowerPointTest/GO/RunTest.sh
@@ -1,17 +1,0 @@
-#!/bin/sh
-TEST_NAME=PDF2WPowerPointTest
-BIN_PATH="../../bin"
-
-if [ ! -d $BIN_PATH ] 
-then
-    mkdir -p $BIN_PATH
-fi
-
-go build -o $BIN_PATH/$TEST_NAME.exe
-
-if [ $? -eq 0 ] 
-then
-    export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:../../../PDFNetC/Lib
-    $BIN_PATH/$TEST_NAME.exe
-fi
-
