--- Samples/PDF2PowerPointTest/PHP/RunTest.bat
@@ -1,6 +1,0 @@
-@echo off
-setlocal
-set TEST_NAME=PDF2PowerPointTest
-set PATH=..\..\..\PDFNetC\Lib;%PATH%
-php.exe %TEST_NAME%.php
-endlocal
