--- Samples/PDF2PowerPointTest/RUBY/RunTest.bat
@@ -1,6 +1,0 @@
-@echo off
-setlocal
-set TEST_NAME=PDF2PowerPointTest
-SET PATH=..\..\..\PDFNetC\Lib;%PATH%
-ruby.exe %TEST_NAME%.rb
-endlocal
