--- Samples/PDF2WordTest/PHP/RunTest.sh
@@ -1,4 +1,0 @@
-#!/bin/sh
-TEST_NAME=PDF2WordTest
-export LD_LIBRARY_PATH=../../../PDFNetC/Lib
-php $TEST_NAME.php
