--- Samples/PDF2WordTest/PYTHON/RunTest.sh
@@ -1,5 +1,0 @@
-#!/bin/sh
-TEST_NAME=PDF2WordTest
-export LD_LIBRARY_PATH=../../../PDFNetC/Lib
-. ../../py_init.sh
-$python_exe -u $TEST_NAME.py
