--- Samples/PDF2WordTest/RUBY/RunTest.sh
@@ -1,4 +1,0 @@
-#!/bin/sh
-TEST_NAME=PDF2WordTest
-export LD_LIBRARY_PATH=../../../PDFNetC/Lib
-ruby $TEST_NAME.rb
